spotify_batch/requirements.txt
pymongo
pyspark
python-dotenv
pandas